- Dodanie pliku funkcje.cpp wraz z nagłówkiem. - Przeniesienie funkcji hamingDistance z main.cpp do funkcje.cpp - Wstępne pomysły, co dalej..
