+++ task_iv_ber/funkcje.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <iostream>
+
+
+uint8_t hammingDistance(uint8_t n1, uint8_t n2); // Funkcja z przyk³adu
